+++ dask/tests/test_base.py
@@ -623,6 +623,16 @@ def test_optimize():
         assert dict(a.dask) == dict(b.dask)
 
 
+# TODO: remove after deprecation cycle of `dask.optimize` module is completed
+def test_optimize_has_deprecated_module_functions_as_attributes():
+    import dask.optimize as deprecated_optimize
+    # Function has method attributes
+    assert dask.optimize.cull is deprecated_optimize.cull
+    assert dask.optimize.inline is deprecated_optimize.inline
+    with pytest.warns(UserWarning):
+        dask.optimize.cull({}, [])
+
+
 def test_default_imports():
     """
     Startup time: `import dask` should not import too many modules.